requirements.txt
pre-commit
flake8
pytest
pytest-cov
GitPython
